requirements.txt
lyricsgenius==3.0.1
python-telegram-bot